Made private `ToJsonArrayWrappingNonObjects` method
